Move external to submodule
