+++ app/build.gradle
@@ -35,6 +35,7 @@ android {
 dependencies {
 
     implementation project(":observer")
+    implementation project(":lastfmdata")
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.7.0'
     implementation 'androidx.appcompat:appcompat:1.4.1'
