+++ app/src/main/java/ayds/lisboa/songinfo/otherdetails/model/OtherDetailsModelInjector.kt
@@ -3,8 +3,8 @@ package ayds.lisboa.songinfo.otherdetails.model
 import android.content.Context
 import ayds.lisboa.songinfo.otherdetails.model.repository.ArtistBiographyRepository
 import ayds.lisboa.songinfo.otherdetails.model.repository.ArtistBiographyRepositoryImpl
-import ayds.lisboa.songinfo.otherdetails.model.repository.external.lastfm.LastFMInjector
-import ayds.lisboa.songinfo.otherdetails.model.repository.external.lastfm.LastFMService
+import ayds.lisboa1.lastfm.LastFMInjector
+import ayds.lisboa1.lastfm.LastFMService
 import ayds.lisboa.songinfo.otherdetails.model.repository.local.lastfm.LastFMLocalStorage
 import ayds.lisboa.songinfo.otherdetails.model.repository.local.lastfm.sqldb.CursorToLastFMArtistBiographyMapperImpl
 import ayds.lisboa.songinfo.otherdetails.model.repository.local.lastfm.sqldb.LastFMLocalStorageImpl
@@ -20,7 +20,7 @@ object OtherDetailsModelInjector {
         val lastFMLocalStorage: LastFMLocalStorage = LastFMLocalStorageImpl(
             otherDetailsView as Context, CursorToLastFMArtistBiographyMapperImpl()
         )
-        val lastFMService: LastFMService = LastFMInjector.lastFMService
+        val lastFMService: ayds.lisboa1.lastfm.LastFMService = ayds.lisboa1.lastfm.LastFMInjector.lastFMService
 
         val repository: ArtistBiographyRepository =
             ArtistBiographyRepositoryImpl(lastFMLocalStorage, lastFMService)
