--- app/src/main/java/ayds/lisboa/songinfo/otherdetails/model/repository/external/lastfm/LastFMInjector.kt
@@ -1,5 +1,0 @@
-package ayds.lisboa.songinfo.otherdetails.model.repository.external.lastfm
-
-object LastFMInjector {
-    val lastFMService : LastFMService = LastFMArtistInjector.lastFMArtistService
-}
