--- app/src/main/java/ayds/lisboa/songinfo/otherdetails/model/repository/external/lastfm/LastFMService.kt
@@ -1,7 +1,0 @@
-package ayds.lisboa.songinfo.otherdetails.model.repository.external.lastfm
-
-import ayds.lisboa.songinfo.otherdetails.model.entities.LastFMArtistBiography
-
-interface LastFMService {
-    fun getArtistBio(artist : String) : LastFMArtistBiography?
-}
